refs #1 when checking for an existing DB, also check whether the DB file is empty
See https://github.com/emmanueltouzery/projectpad2/issues/1
if you start the app, and close the login screen without
unlocking the DB, we leave a DB file of zero bytes, and at
next startup we ask you for the unlock password, we don't
anymore ask you for a confirm password, because we assume
there's already a DB around => check that the db file is
present AND not empty.
if reading the file length fails, assume a non-empty file.