fix export on flatpak with file not ending in 7z
When opening the file picker dialog to save a file,
you pick the folder and enter a file name.
Flatpak gives access to folder/file, and ONLY THAT. Nothing else.
If the filename doesn't end in .7z, I'd add the extension,
but flatpak would NOT grant me access to folder/file.7z!
So the export would fail.
If in the export dialog the user properly types .7z in the filename, everything works.
Unfortunately the file chooser extension filter doesn't work from flatpak...