more emphasized mode switch button
spent quite some time thinking about how to display this, including
whether to put it in the header bar, in the contents, have one button
per section and so on. ultimately, i think this is the clearest, least
busy interface and i think it's not too out of place with gnome.